Clarify invite link to Discord on reference page
When you look up "ProgrammingLanguages Discord", this page is the first result. But a big list of rules isn't a good introduction, and it's not obvious from the CSS styles that the big purple title is an invite link (or that the site name is a link). Maybe that's a usability issue that we didn't notice?